refactor: Tacticクラスを個別ファイルに分割
skill_tactic.hpp に集約されていた複数のTacticクラスを、
他のTacticと同様に1クラス1ファイルの構成に変更しました。

変更内容:
- GoalieSkillTactic を goalie_skill_tactic.hpp/.cpp に分割
- BallPlacementSkillTactic を ball_placement_skill_tactic.hpp/.cpp に分割
- SubAttackerSkillTactic を sub_attacker_skill_tactic.hpp/.cpp に分割
- SimpleKickOffSkillTactic を simple_kick_off_skill_tactic.hpp/.cpp に分割
- BallNearByPositionerSkillTactic を ball_near_by_positioner_skill_tactic.hpp/.cpp に分割
- PlacementTargetNearByPositionerSkillTactic を placement_target_near_by_positioner_skill_tactic.hpp/.cpp に分割
- tactic_factory.cpp のインクルードを更新
- skill_tactic.hpp と skill_planners.cpp を削除

この変更により、コードの可読性と保守性が向上します。
